+++ lib/features/course/presentation/widgets/first_part_of_row_widget.dart
