+++ lib/features/course/presentation/widgets/second_part_of_row_widget.dart
