+++ lib/main.dart
@@ -4,7 +4,7 @@ import 'package:classroom_app/core/pages/main_page.dart';
 import 'package:classroom_app/features/home/presentation/pages/home_page.dart';
 
 import 'features/auth/presentation/pages/login_page.dart';
-import 'features/classroom/presentation/pages/classroom_page.dart';
+import 'features/course/presentation/pages/course_page.dart';
 import 'features/landing/presentation/pages/landing_page.dart';
 import 'features/user/presentation/pages/user_page.dart';
 
@@ -28,7 +28,7 @@ class MyApp extends StatelessWidget {
         UserPage.routeName: (_) => const UserPage(),
         MainPage.routeName: (_) => const MainPage(),
         LogInPage.routeName: (_) => const LogInPage(),
-        ClassroomPage.routeName: (_) => const ClassroomPage(),
+        CoursePage.routeName: (_) => const CoursePage(),
       },
       theme: ThemeData(
         useMaterial3: true,
